refactor: 계산 로직 hook으로 분리
